+++ SimTracker/SiStripDigitizer/plugins/DigiSimLinkPileUpSignals.cc
@@ -6,7 +6,7 @@ void DigiSimLinkPileUpSignals::resetLink(){
   theCounterMapLink.clear();
 }
 
-void DigiSimLinkPileUpSignals::add(const std::vector<double>& locAmpl,
+void DigiSimLinkPileUpSignals::add(const std::vector<float>& locAmpl,
 			  const size_t& firstChannelWithSignal, const size_t& lastChannelWithSignal,
 			  const PSimHit* hit,const int& counter){
   for (size_t iChannel=firstChannelWithSignal; iChannel<lastChannelWithSignal; ++iChannel) {
