+++ SimTracker/SiStripDigitizer/plugins/SiGaussianTailNoiseAdder.cc
@@ -15,7 +15,7 @@ SiGaussianTailNoiseAdder::SiGaussianTailNoiseAdder(float th,CLHEP::HepRandomEngi
 SiGaussianTailNoiseAdder::~SiGaussianTailNoiseAdder(){
 }
 
-void SiGaussianTailNoiseAdder::addNoise(std::vector<double> &in,
+void SiGaussianTailNoiseAdder::addNoise(std::vector<float> &in,
 					size_t& minChannel, size_t& maxChannel,
 					int numStrips, float noiseRMS) const {
  
@@ -38,7 +38,7 @@ void SiGaussianTailNoiseAdder::addNoise(std::vector<double> &in,
   }
 }
 
-void SiGaussianTailNoiseAdder::addNoiseVR(std::vector<double> &in, std::vector<float> &noiseRMS) const {
+void SiGaussianTailNoiseAdder::addNoiseVR(std::vector<float> &in, std::vector<float> &noiseRMS) const {
   // Add noise
   // Full Gaussian noise is added everywhere
   for (size_t iChannel=0; iChannel!=in.size(); iChannel++) {
@@ -46,13 +46,13 @@ void SiGaussianTailNoiseAdder::addNoiseVR(std::vector<double> &in, std::vector<f
   }
 }
 
-void SiGaussianTailNoiseAdder::addPedestals(std::vector<double> &in,std::vector<float> & ped) const {
+void SiGaussianTailNoiseAdder::addPedestals(std::vector<float> &in,std::vector<float> & ped) const {
 	for (size_t iChannel=0; iChannel!=in.size(); iChannel++) {
       if(ped[iChannel]>0.) in[iChannel] += ped[iChannel];      
     }
 }
 
-void SiGaussianTailNoiseAdder::addCMNoise(std::vector<double> &in, float cmnRMS, std::vector<bool> &badChannels) const {
+void SiGaussianTailNoiseAdder::addCMNoise(std::vector<float> &in, float cmnRMS, std::vector<bool> &badChannels) const {
   int nAPVs = in.size()/128;
   std::vector<float> CMNv;
   for(int APVn =0; APVn < nAPVs; ++APVn) CMNv.push_back(gaussDistribution->fire(0.,cmnRMS));
@@ -61,7 +61,7 @@ void SiGaussianTailNoiseAdder::addCMNoise(std::vector<double> &in, float cmnRMS,
   }
 }
 
-void SiGaussianTailNoiseAdder::addBaselineShift(std::vector<double> &in, std::vector<bool> &badChannels) const {
+void SiGaussianTailNoiseAdder::addBaselineShift(std::vector<float> &in, std::vector<bool> &badChannels) const {
   size_t nAPVs = in.size()/128;
   std::vector<float> vShift;
   double apvCharge, apvMult;
