+++ SimTracker/SiStripDigitizer/plugins/SiHitDigitizer.cc
@@ -30,7 +30,7 @@ SiHitDigitizer::~SiHitDigitizer(){
 
 void 
 SiHitDigitizer::processHit(const PSimHit* hit, const StripGeomDetUnit& det, GlobalVector bfield,float langle,
-			   std::vector<double>& locAmpl, size_t& firstChannelWithSignal, size_t& lastChannelWithSignal,
+			   std::vector<float>& locAmpl, size_t& firstChannelWithSignal, size_t& lastChannelWithSignal,
 			   const TrackerTopology *tTopo){
   
   // Compute the drift direction for this det
