+++ SimTracker/SiStripDigitizer/plugins/SiInduceChargeOnStrips.h
@@ -20,7 +20,7 @@ public:
   
   virtual ~SiInduceChargeOnStrips() { }
   virtual void induce(const SiChargeCollectionDrifter::collection_type&, const StripGeomDetUnit&, 
-		      std::vector<double>&, size_t&, size_t&,
+		      std::vector<float>&, size_t&, size_t&,
 		      const TrackerTopology *tTopo) const=0;
 
 };
