+++ SimTracker/SiStripDigitizer/plugins/SiPileUpSignals.cc
@@ -5,7 +5,7 @@ void SiPileUpSignals::resetSignals(){
   signal_.clear();
 }
 
-void SiPileUpSignals::add(uint32_t detID, const std::vector<double>& locAmpl,
+void SiPileUpSignals::add(uint32_t detID, const std::vector<float>& locAmpl,
                           const size_t& firstChannelWithSignal, const size_t& lastChannelWithSignal) {
   SignalMapType& theSignal = signal_[detID];
   for (size_t iChannel=firstChannelWithSignal; iChannel<lastChannelWithSignal; ++iChannel) {
