+++ SimTracker/SiStripDigitizer/plugins/SiTrivialInduceChargeOnStrips.cc
@@ -81,7 +81,7 @@ void
 SiTrivialInduceChargeOnStrips::
 induce(const SiChargeCollectionDrifter::collection_type& collection_points, 
        const StripGeomDetUnit& det, 
-       std::vector<double>& localAmplitudes, 
+       std::vector<float>& localAmplitudes, 
        size_t& recordMinAffectedStrip, 
        size_t& recordMaxAffectedStrip,
        const TrackerTopology *tTopo) const {
